Add a simple role to add a grain to salt
The idea is to be able to know from within salt that a server
is now managed by ansible and that salt shouldn't touch it.

Main use is to prevent openssh and freeipa integration to be broken.

Since salt requires to manage the whole file (file.line is in 2015.8, not in
EPEL 7), the openssh config lost the freeipa integration for ssh key, which is
more annoying than planned. So I want to switch that to ansible, but
I do not want to start to exclude server one by one in top.sls, so this
is a automated way of doing it.