Use a better scheme for generating password
As the datetime could be preditcted (since it will be just before
the commit appear on github, a smart attacker could reduce the date
to a 10 seconds timeframe, which mean around 10 millions of possible
date to pass to md5. That's around 27h of bruteforce if doing 100
attempts per second.

Adding shuffle|join do change completely that, since suddenly, the
string is generated is much more random (albeit I wasn't able to
verify how much, but it was more than 10**20 on my estimation), and
that's anyway reduced to the 128 bits of the md5 sum, which should be
enough (equivalent to a 16 chars password, if I am not wrong).