Fix the long standing issue of proxy and encoded slash
So mod_rewrite was not suitable for gerrit proxy, because
of a issue with encoded slash, and the fact it do decode
them. So we have to setup the proxy outside of the httpd role for now
 But we did switch to mod_rewrite because it was simpler to filter
call for letsencrypt and fix various long standing bugs.
It turn out that you can also filter URL on mod_proxy, feature
that I used now for that snipet.