Add changeCleanup, remove anon_git, and fix scopes
changeCleanup is a conversation we've had for some time. This is testing
out how it works and making sure it works. Once it's tested, we may want
to change the interval to once a week rather than hourly.

anon_git should no longer be advertised because we really want to switch
over from that.

the scopes issue was working with an earlier version of Gerrit, but
newer Gerrits do not like it anymore. It causes straight up Server Error
for new users.