Do not redeploy syslog on supercolony
Due to likely some bug in rsyslog, the queue get full when rsyslog
can't export logs to remote server. After some outage on freeipa that
I still didn't fix due to hackweek, OSS-EU and general work.

So in the mean time, we fix by changing syslog configuration and
by not applying rsyslog change on it to revert the fix.