Add the last part of the proxies in HA
It use nginx to be able to use the default vhost,
matching unnamed vhost for the http01 acme negociation,
and the openssl_* modules for creating the certificate
at the right place to be replcaited thanks to gluster.

It also force SSL as a way to simplify the code and avoid
corner case.