Bundle package installation as 1 transaction
In the past, ansible was bundling all of them at once, but this
changed. Thanks to @Pilou- and @nigelbabu for the idea and the
way to improve that. Measurement show this save around 3 to 4
minutes per builder, so likely saving half a hour since we have
around 60 builders, and running on 10 of them at the same time.