Properly fix the issue of ci-logs.
For some reason, the variable inheritence changed again in Ansible, and
that has the side effect of breaking the playbook. As I have better us of my time
than trying to track down again some breakage, I work around the problem
by rewriting the playbook until next time. Maybe it wasn't finished or not
supposed to work that way.