Verify certificate expiration, and use SNI
By default, check_http do not seems to use SNI, so it likely verify
only the first ssl vhost. This is rather suboptimal, even
if I guess that in practice, the http check would catch any issue.

It also verify the certificate is going to expire in less than 1 week and
our certificate should renew 10 days before the expiration.